# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Store {
  id: number;
  name: string;
  city: string;
  state: string;
  channel: string;
  createdAt: string;
}

export interface StoreInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  channel: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
}

export type VisitItemShelfCondition = typeof VisitItemShelfCondition[keyof typeof VisitItemShelfCondition];


export const VisitItemShelfCondition = {
  good: 'good',
  regular: 'regular',
  bad: 'bad',
} as const;

export interface VisitItem {
  productId: number;
  inStock: boolean;
  /** @nullable */
  price: number | null;
  shelfCondition: VisitItemShelfCondition;
  /** @nullable */
  notes?: string | null;
}

export type VisitItemDetailShelfCondition = typeof VisitItemDetailShelfCondition[keyof typeof VisitItemDetailShelfCondition];


export const VisitItemDetailShelfCondition = {
  good: 'good',
  regular: 'regular',
  bad: 'bad',
} as const;

export interface VisitItemDetail {
  id: number;
  visitId: number;
  productId: number;
  productName: string;
  productCategory: string;
  inStock: boolean;
  /** @nullable */
  price: number | null;
  shelfCondition: VisitItemDetailShelfCondition;
  /** @nullable */
  notes: string | null;
}

export interface VisitInput {
  storeId: number;
  visitedAt: string;
  /** @nullable */
  notes?: string | null;
  items: VisitItem[];
}

export interface Visit {
  id: number;
  storeId: number;
  storeName: string;
  storeCity: string;
  storeState: string;
  storeChannel: string;
  visitedAt: string;
  /** @nullable */
  notes: string | null;
  items: VisitItemDetail[];
}

export interface DashboardSummary {
  totalStores: number;
  totalProducts: number;
  totalVisits: number;
  outOfStockCount: number;
  poorShelfCount: number;
  visitsThisWeek: number;
}

export interface OutOfStockItem {
  storeId: number;
  storeName: string;
  storeCity: string;
  storeChannel: string;
  productId: number;
  productName: string;
  productCategory: string;
  visitId: number;
  visitedAt: string;
}

export interface PoorShelfItem {
  storeId: number;
  storeName: string;
  storeCity: string;
  storeChannel: string;
  productId: number;
  productName: string;
  shelfCondition: string;
  visitId: number;
  visitedAt: string;
}

export interface RecentVisit {
  id: number;
  storeId: number;
  storeName: string;
  storeCity: string;
  storeChannel: string;
  visitedAt: string;
  /** @nullable */
  notes?: string | null;
  itemCount: number;
  outOfStockCount: number;
  poorShelfCount: number;
}