# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalStores: number;
  totalProducts: number;
  totalVisits: number;
  outOfStockCount: number;
  poorShelfCount: number;
  visitsThisWeek: number;
}