# lib/api-zod/src/generated/types/outOfStockItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */

export interface OutOfStockItem {
  storeId: number;
  storeName: string;
  storeCity: string;
  storeChannel: string;
  productId: number;
  productName: string;
  productCategory: string;
  visitId: number;
  visitedAt: string;
}