# lib/api-zod/src/generated/types/poorShelfItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */

export interface PoorShelfItem {
  storeId: number;
  storeName: string;
  storeCity: string;
  storeChannel: string;
  productId: number;
  productName: string;
  shelfCondition: string;
  visitId: number;
  visitedAt: string;
}