# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
}