# lib/api-zod/src/generated/types/recentVisit.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */

export interface RecentVisit {
  id: number;
  storeId: number;
  storeName: string;
  storeCity: string;
  storeChannel: string;
  visitedAt: string;
  /** @nullable */
  notes?: string | null;
  itemCount: number;
  outOfStockCount: number;
  poorShelfCount: number;
}