# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */
import type { VisitItemDetail } from './visitItemDetail';

export interface Visit {
  id: number;
  storeId: number;
  storeName: string;
  storeCity: string;
  storeState: string;
  storeChannel: string;
  visitedAt: string;
  /** @nullable */
  notes: string | null;
  items: VisitItemDetail[];
}