# lib/api-zod/src/generated/types/visitInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */
import type { VisitItem } from './visitItem';

export interface VisitInput {
  storeId: number;
  visitedAt: string;
  /** @nullable */
  notes?: string | null;
  items: VisitItem[];
}