# lib/api-zod/src/generated/types/visitItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */
import type { VisitItemShelfCondition } from './visitItemShelfCondition';

export interface VisitItem {
  productId: number;
  inStock: boolean;
  /** @nullable */
  price: number | null;
  shelfCondition: VisitItemShelfCondition;
  /** @nullable */
  notes?: string | null;
}