# lib/api-zod/src/generated/types/visitItemDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trade Marketing API
 * OpenAPI spec version: 0.1.0
 */
import type { VisitItemDetailShelfCondition } from './visitItemDetailShelfCondition';

export interface VisitItemDetail {
  id: number;
  visitId: number;
  productId: number;
  productName: string;
  productCategory: string;
  inStock: boolean;
  /** @nullable */
  price: number | null;
  shelfCondition: VisitItemDetailShelfCondition;
  /** @nullable */
  notes: string | null;
}